src/test/java/com/deerhunter/Topic050Test.java
package com.deerhunter;

import com.deerhunter.topic.Topic050;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

/**
 * Oops, forgot to write comment. Good luck, bro.
 * <p>
 * Copyright (c) 2019, deerhunter0837@gmail.com All Rights Reserved.
 *
 * @author xuejunc
 * @date 2020-03-07
 */
class Topic050Test {
    @Test
    void solution1() {
        Topic050.Solution1 instance = new Topic050.Solution1();
        assertEquals(1024.0, instance.myPow(2.0, 10));
        assertEquals(9.26100, instance.myPow(2.10000, 3));
        assertEquals(0.25000, instance.myPow(2.0, -2));
    }

    @Test
    void solution2() {
        Topic050.Solution2 instance = new Topic050.Solution2();
        assertEquals(1024.0, instance.myPow(2.0, 10));
        assertEquals(9.26100, instance.myPow(2.10000, 3));
        assertEquals(0.25000, instance.myPow(2.0, -2));
    }

    @Test
    void solution3() {
        Topic050.Solution3 instance = new Topic050.Solution3();
        assertEquals(1024.0, instance.myPow(2.0, 10));
        assertEquals(9.26100, instance.myPow(2.10000, 3));
        assertEquals(0.25000, instance.myPow(2.0, -2));
    }

    @Test
    void solution4() {
        Topic050.Solution4 instance = new Topic050.Solution4();
        assertEquals(1024.0, instance.myPow(2.0, 10));
        assertEquals(9.26100, instance.myPow(2.10000, 3));
        assertEquals(0.25000, instance.myPow(2.0, -2));
    }
}
